+++ sugarcube.d.ts
@@ -12,7 +12,6 @@ declare module 'twine-sugarcube' {
     ignoreGender?: boolean
     showTutorial?: boolean
     showBiomeGeneration?: boolean
-    showSliders?: boolean
     displayTwoColumns?: boolean
     hideAds?: boolean
     disableAnalytics?: boolean